+++ Back-end/api-module/src/main/java/com/ico/api/dto/coolsms/PhoneNumReqDto.java
@@ -1,0 +1,17 @@
+package com.ico.api.dto.coolsms;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+
+/**
+ * 휴대폰 번호 받는 Dto
+ */
+@Getter
+@NoArgsConstructor
+public class PhoneNumReqDto {
+
+    @NotBlank(message = "107")
+    private String phoneNum;
+}
